set up inheritance structure for all players
